refactor: 修改地址引用方式
