+++ src/components/buys/AddressCard.vue
@@ -1,17 +1,6 @@
 <script lang="ts" setup>
-import areadata from '@/assets/json/division.json'
-
 const { nowAddress } = storeToRefs(useAddressStore())
 // 通过code获取省市区
-
-const getPcaDetails = (pcaCode: string[]) => {
-  const [p, c, a] = pcaCode
-  const province = areadata.find(item => item.value === p)
-  const city = province?.children.find(item => item.value === c)
-  const area = city?.children.find(item => item.value === a)
-
-  return `${province?.label} ${city?.label} ${area?.label}`
-}
 onShow(() => {
 
 })
